heySeven/requirements.txt
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn==0.30.6
pydantic==2.9.2
langgraph==0.2.28
langchain-openai==0.2.3
langchain-core==0.3.12
pytest==8.3.3
httpx==0.27.2